Reorganise docs and centralise Sprite types
Restructure docs/ into subdirectories (business/, engineering/, concepts/, apps/) with README index files and updated cross-links. Remove legacy README_leida.md. Add Sprite documentation at docs/apps/flash/sprite.md.

Consolidate Sprite types from local types.ts into the package-root types.d.ts, removing the now-redundant Sprite/types.ts file. Update all internal imports accordingly.

Refactor the flash demo to show a single large sprite with Prev/Next direction controls instead of a grid layout.
