+++ docs/apps/README.md
@@ -1,0 +1,6 @@
+# App Docs
+
+This folder contains app-specific documentation.
+
+- [Admin](./admin/admin.md)
+- [Flash Sprite](./flash/sprite.md)
