+++ docs/apps/admin/cartridges/nx-admin/hooks.md
