+++ docs/apps/admin/cartridges/nx-admin/pwa.md
