+++ docs/business/README.md
@@ -1,0 +1,8 @@
+# Business Docs
+
+This folder contains business-facing documentation for strategy, delivery, ownership, and investment framing.
+
+- [Executive Overview](./executive-overview.md)
+- [Owner Guide](./owner-guide.md)
+- [Investor Overview](./investor-overview.md)
+- [Operations and Delivery](./operations-and-delivery.md)
