+++ docs/business/owner-guide.md
@@ -148,4 +148,4 @@ An accurate owner-level statement would be:
 - [Executive Overview](./executive-overview.md)
 - [Investor Overview](./investor-overview.md)
 - [Operations and Delivery](./operations-and-delivery.md)
-- [Project Evolution](./project-evolution.md)
+- [Project Evolution](../concepts/project-evolution.md)
