+++ docs/concepts/README.md
@@ -1,0 +1,5 @@
+# Concepts Docs
+
+This folder contains conceptual and historical context for how the platform has evolved.
+
+- [Project Evolution](./project-evolution.md)
