+++ docs/engineering/README.md
@@ -1,0 +1,10 @@
+# Engineering Docs
+
+This folder contains technical documentation for architecture, implementation, and developer workflows.
+
+- [Developer Guide](./developer-guide.md)
+- [Tech Stack](./techstack.md)
+- [Apps and Packages](./apps-packages.md)
+- [Testing Strategy](./testing-strategy.md)
+- [Storybook](./storybook.md)
+- [Developer](./developer.md)
