+++ docs/engineering/apps-packages.md
@@ -18,12 +18,12 @@ This monorepo is organized into two main kinds of workspaces: apps and packages.
 ## Apps
 Apps are deployable products or experiences. They contain the user-facing application code and are the places where end-to-end features are assembled.
 
-In this repository, the main app is located under [apps/nx](../apps/nx), and it represents the primary web experience.
+In this repository, the primary app surfaces are [apps/www](../../apps/www) for the public web experience and [apps/admin](../../apps/admin) for internal operations.
 
 ## Packages
 Packages are reusable building blocks shared across apps or internal tooling. They are typically smaller, focused modules that can be imported by multiple parts of the monorepo.
 
-Examples in this repository include shared libraries under [packages](../packages), such as the Firebase, Flash, and Shortcodes packages.
+Examples in this repository include shared libraries under [packages](../../packages), such as [packages/design-system](../../packages/design-system), [packages/uberedux](../../packages/uberedux), and utility packages like [packages/firebase](../../packages/firebase).
 
 ## How to think about them
 - Use apps for product experiences and deployment targets.
