+++ packages/flash/src/MovieClips/Sprite/index.ts
@@ -1,6 +1,6 @@
 import Sprite from './Sprite';
 
-export type { Direction, SpriteState, SpriteProps, SpritePalette } from './types';
+export type { Direction, SpriteState, SpriteProps, SpritePalette } from '../../../types';
 export { getDirection } from './direction';
 export { Sprite };
 export default Sprite;
